# ordering/internal/ports/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ordering Service API",
    "version": "1.0",
    "contact": {
      "name": "Ordering service",
      "url": "https://github.com/rezaAmiri123/microservice"
    }
  },
  "tags": [
    {
      "name": "OrderingService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/api/ordering/cancel_order": {
      "post": {
        "summary": "Cancel an order",
        "description": "Use this API to cancel an order",
        "operationId": "OrderingService_CancelOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbCancelOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderingpbCancelOrderRequest"
            }
          }
        ],
        "tags": [
          "OrderingService"
        ]
      }
    },
    "/v1/api/ordering/complete_order": {
      "post": {
        "summary": "Complete an order",
        "description": "Use this API to complete an order",
        "operationId": "OrderingService_CompleteOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbCompleteOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderingpbCompleteOrderRequest"
            }
          }
        ],
        "tags": [
          "OrderingService"
        ]
      }
    },
    "/v1/api/ordering/create_order": {
      "post": {
        "summary": "Create new order",
        "description": "Use this API to create a new order",
        "operationId": "OrderingService_CreateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbCreateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderingpbCreateOrderRequest"
            }
          }
        ],
        "tags": [
          "OrderingService"
        ]
      }
    },
    "/v1/api/ordering/get_order": {
      "post": {
        "summary": "Get an order",
        "description": "Use this API to get an order",
        "operationId": "OrderingService_GetOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbGetOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderingpbGetOrderRequest"
            }
          }
        ],
        "tags": [
          "OrderingService"
        ]
      }
    },
    "/v1/api/ordering/ready_order": {
      "post": {
        "summary": "Ready order",
        "description": "Use this API to make ready an order",
        "operationId": "OrderingService_ReadyOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderingpbReadyOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderingpbReadyOrderRequest"
            }
          }
        ],
        "tags": [
          "OrderingService"
        ]
      }
    }
  },
  "definitions": {
    "orderingpbCancelOrderRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "orderingpbCancelOrderResponse": {
      "type": "object"
    },
    "orderingpbCompleteOrderRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "invoiceId": {
          "type": "string"
        }
      }
    },
    "orderingpbCompleteOrderResponse": {
      "type": "object"
    },
    "orderingpbCreateOrderRequest": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/orderingpbItem"
          }
        },
        "userId": {
          "type": "string"
        },
        "paymentId": {
          "type": "string"
        }
      }
    },
    "orderingpbCreateOrderResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "orderingpbGetOrderRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "orderingpbGetOrderResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/orderingpbOrder"
        }
      }
    },
    "orderingpbItem": {
      "type": "object",
      "properties": {
        "storeId": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "storeName": {
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "orderingpbOrder": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "paymentId": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/orderingpbItem"
          }
        },
        "status": {
          "type": "string"
        }
      }
    },
    "orderingpbReadyOrderRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "orderingpbReadyOrderResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}